[#252] Split MpeChannelAllocator's expression, identity and state types out of its file
MpeChannelAllocator.scala held 18 top-level declarations in 929 lines, of
which the allocator and its companion were only two. Move the Expression
Value model, the note identity and the allocator's internal state classes
into their own files, leaving the allocator file with the allocation
algorithm and the result vocabulary it produces (642 lines):

- MpeExpression.scala — MpeExpression + companion, MutableMpeExpression,
  ImmutableMpeExpression + companion, MpeExpressionUpdate + companion,
  MpeChannelExpressionUpdate;
- MpeNoteIdentity.scala — MpeNoteIdentity;
- MpeChannelState.scala — MpeNoteState, MpeChannelState.

Restrict every one of the 18 declarations to private[tuner], the allocator
and its companion included. They are implementation detail shared between
MpeTuner and MpeChannelAllocator, not API for the rest of the module: all
eleven allocator references in MpeTuner.scala sit inside private members,
and nothing outside the tuner module refers to any of these types. The
three formerly top-level `private` classes get the explicit modifier too,
since Scala 3 lifts top-level definitions into a synthetic $package object
where the file-scoped reading of bare `private` is arguable. Companion
objects do not inherit the modifier from their class, so each carries its
own.

Also drop the meaningless `case` modifier from ImmutableMpeExpression's
companion, whose declaration line was edited anyway.

No behaviour change: the moved code is byte-identical apart from the
visibility modifiers and the parameter-list realignment they force. The
existing tests pass unchanged.
